requirements.txt
yfinance
pandas
numpy
statsmodels
scikit-learn
matplotlib
scipy
gymnasium
stable-baselines3
alpaca-py
python-dotenv
